feat: Add base container functions - Implement BaseAlpine, BaseUbuntu, BaseBusyBox, and BaseWolfi methods to create base containers with different Linux distributions
