+++ module-template/container_base.go
